# public/api-documentation/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel Swagger Documentation",
        "contact": {
            "email": "peacevan@hotmail.com"
        },
        "version": "1.5.0"
    },
    "paths": {
        "/api/cliente": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Listar Clientes",
                "description": "Retorna todos os Clientes",
                "operationId": "3523670cf42771e8e78e1a3d40f3ff61",
                "responses": {
                    "200": {
                        "description": "Clientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "5"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Cliente 1 Teste"
                                        },
                                        "cnpj": {
                                            "type": "string",
                                            "example": "000220007878"
                                        },
                                        "data_fundacao": {
                                            "type": "string",
                                            "example": "17/05/2010"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campo Incorreto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "esse nome já está sendo usado"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Criar um novo Cliente",
                "description": "esse endpoint cria um novo Cliente",
                "operationId": "88727081caadff2d758122f42b2b1eec",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "id_grupo"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Ivan Amado"
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "example": "1212122122122"
                                    },
                                    "data_fundacao": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "17/05/2010"
                                    },
                                    "id_grupo": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Grupo Cliado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "group created successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Incorrect fields",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name has already been taken. (and 2 more errors)"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cliente/{id}": {
            "put": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Mudar dados do cliente",
                "description": "esse endpoint muda os dados do cliente",
                "operationId": "7f8589f93cf4cbb63b728d2e8dd3d973",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "id_grupo"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Ivan Amado"
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "example": "1212122122122"
                                    },
                                    "data_fundacao": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "17/05/2010"
                                    },
                                    "id_grupo": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente atualizado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Incorrect fields",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Esse Nome já foi cadastrado"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Remove um Cliente",
                "description": "Esse Endpoint remove um Gerente.",
                "operationId": "d01497479a258592c5d5fa8778b537eb",
                "responses": {
                    "204": {
                        "description": "Remove um cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente removido com sucesso !"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Incorrect fields",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "nome do Cliente é obrigatório."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "Acess token obtido na autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}